refactor: factor various fs utilities out
