+++ internal/promotion/runner/builtin/kustomize_builder_test.go
@@ -13,7 +13,7 @@ import (
 	"helm.sh/helm/v3/pkg/repo"
 
 	kargoapi "github.com/akuity/kargo/api/v1alpha1"
-	"github.com/akuity/kargo/internal/fs"
+	"github.com/akuity/kargo/internal/io/fs"
 	"github.com/akuity/kargo/pkg/promotion"
 	"github.com/akuity/kargo/pkg/x/promotion/runner/builtin"
 )
